backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
python-multipart
python-jose[cryptography]
passlib[bcrypt]
langchain
langchain-community
faiss-cpu
sentence-transformers
transformers
torch
accelerate
bitsandbytes
python-dotenv
asyncpg
supabase
pypdf
